add strapp, strprep & update pow, itoa_base

+++ ft_itoa_base.c
@@ -1,12 +1,12 @@
 #include <stdlib.h>
 
-char	get_char(long value, int base)
+char	get_char(long long value, int base)
 {
 	char *str = "0123456789ABCDEF";
 	return str[value % base];
 }
 
-void get_value(char *str, long value, int base, int *pos)
+void get_value(char *str, long long value, int base, int *pos)
 {
 	if(value >= base)
 		get_value(str, value / base, base, pos);
@@ -14,11 +14,11 @@ void get_value(char *str, long value, int base, int *pos)
 	*pos += 1;
 }
 
-char	*ft_itoa_base(int value, int base)
+char	*ft_itoa_base(long long value, int base)
 {
 	char *result;
 	int i;
-	long tmp;
+	long long tmp;
 
 	i = 0;
 	result = (char*)malloc(sizeof(char) * (25));
